pyproject.toml
[project]
name = "central-knowledge-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "dvc>=3.59.2",
    "dvc-s3>=3.2.0",
    "google-adk>=0.5.0",
    "lightrag-hku[api]>=1.3.6",
    "litellm>=1.69.0",
    "loguru>=0.7.3",
    "openai>=1.57.0",
    "pip>=25.1.1",
    "pytest>=8.3.5",
    "gitingest>=0.1.4",
    "nest-asyncio>=1.6.0",
    "pytest-cov>=6.1.1",
    "jira>=3.8.0",
    "pytest-pythonpath>=0.7.3",
    "tqdm>=4.66.4",
    "fastmcp>=2.5.2",
    "requests>=2.31.0",
    "graphiti-core>=0.11.6",
    "tenacity>=9.1.2",
    "aiofiles>=24.1.0",
    "neo4j>=5.28.1",
    "python-dotenv>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]
